Update pis-functions.php to avoid erroring
When dealing with taxonomies, PIS expects to receive a valid set of taxonomy details to work with. However, for sites which use CPT-onomies, there is a long outstanding bug which populates term cache in WP 4.7.2 and above with NULLs, hence when PIS touches such a Post, as it wasn't expecting a NULL it would cause to error out and the page with the post on would not get properly generated. An additional check for a WP_Error return value is now added to handle this circumstance.